fix: remove more button from saved movies page
