+++ src/components/MoviesCardList/MoviesCardList.css
@@ -18,37 +18,6 @@
   padding: 0 0 80px 0;
 }
 
-.movies-card-list__more-button-container {
-  min-height: 10px;
-  
-  padding: 0;
-}
-
-.movies-card-list__more-button {
-  width: 100%;
-  
-  color: #fff;
-  font-size: 12px;
-  font-weight: 500;
-  line-height: normal;
-  cursor: pointer;
-  background-color: #2f2f2f;
-  
-  opacity: 1;
-  
-  border: none;
-  border-radius: 6px;
-  padding: 10px 0;
-  
-  transition: opacity .4s ease-out;
-}
-
-.movies-card-list__more-button:hover {
-  opacity: .5;
-  
-  transition: opacity .2s ease-in;
-}
-
 @media (min-width: 480px) {
   .movies-card-list__container {
     grid-template-columns: repeat(auto-fill, minmax(348px, auto));
@@ -66,10 +35,6 @@
   
   .movies-card-list__container {
     grid-template-columns: repeat(auto-fill, minmax(360px, auto));
-  }
-  
-  .movies-card-list__more-button-container {
-    padding-bottom: 70px;
   }
 }
